sf::Vector impl

+++ inc/gameobject/property/mood.h
@@ -5,10 +5,10 @@ namespace Property
 {
     struct Mood
     {
-        double stresslevel;
-        double angrylevel;
-        double inLove;
-        double druglevel;
+        float stresslevel;
+        float angrylevel;
+        float inLove;
+        float druglevel;
     };
 }
 #endif // MOOD_H
